+++ contracts/mocks/account/paymaster/PaymasterNFTMock.sol
@@ -3,12 +3,10 @@
 pragma solidity ^0.8.20;
 
 import {Ownable} from "@openzeppelin/contracts/access/Ownable.sol";
-import {PackedUserOperation} from "@openzeppelin/contracts/interfaces/draft-IERC4337.sol";
-import {ERC4337Utils} from "@openzeppelin/contracts/account/utils/draft-ERC4337Utils.sol";
-import {PaymasterSigner} from "../../../account/paymaster/PaymasterSigner.sol";
-import {SignerECDSA} from "../../../utils/cryptography/SignerECDSA.sol";
+import {ERC4337Utils, PackedUserOperation} from "@openzeppelin/contracts/account/utils/draft-ERC4337Utils.sol";
+import {PaymasterNFT} from "../../../account/paymaster/PaymasterNFT.sol";
 
-abstract contract PaymasterCoreContextNoPostOpMock is PaymasterSigner, SignerECDSA, Ownable {
+abstract contract PaymasterNFTContextNoPostOpMock is PaymasterNFT, Ownable {
     using ERC4337Utils for *;
 
     function _validatePaymasterUserOp(
@@ -16,16 +14,16 @@ abstract contract PaymasterCoreContextNoPostOpMock is PaymasterSigner, SignerECD
         bytes32 userOpHash,
         uint256 requiredPreFund
     ) internal override returns (bytes memory context, uint256 validationData) {
-        // use the userOp's paymasterData as context;
-        context = userOp.paymasterData();
-        // super call (PaymasterSigner + SignerECDSA) for the validation data
+        // use the userOp's callData as context;
+        context = userOp.callData;
+        // super call (PaymasterNFT) for the validation data
         (, validationData) = super._validatePaymasterUserOp(userOp, userOpHash, requiredPreFund);
     }
 
     function _authorizeWithdraw() internal override onlyOwner {}
 }
 
-abstract contract PaymasterCoreMock is PaymasterCoreContextNoPostOpMock {
+abstract contract PaymasterNFTMock is PaymasterNFTContextNoPostOpMock {
     event PaymasterDataPostOp(bytes paymasterData);
 
     function _postOp(