+++ test/account/paymaster/PaymasterNFT.test.js
@@ -2,23 +2,23 @@ const { ethers, entrypoint } = require('hardhat');
 const { loadFixture } = require('@nomicfoundation/hardhat-network-helpers');
 
 const { getDomain } = require('@openzeppelin/contracts/test/helpers/eip712');
-const { PackedUserOperation, UserOperationRequest } = require('../../helpers/eip712-types');
+const { PackedUserOperation } = require('../../helpers/eip712-types');
 const { ERC4337Helper } = require('../../helpers/erc4337');
 
 const { shouldBehaveLikePaymaster } = require('./Paymaster.behavior');
 
 for (const [name, opts] of Object.entries({
-  PaymasterCore: { postOp: true },
-  PaymasterCoreContextNoPostOp: { postOp: false },
+  PaymasterNFT: { postOp: true, timeRange: false },
+  PaymasterNFTContextNoPostOp: { postOp: false, timeRange: false },
 })) {
   async function fixture() {
     // EOAs and environment
     const [admin, receiver, other] = await ethers.getSigners();
     const target = await ethers.deployContract('CallReceiverMockExtended');
+    const token = await ethers.deployContract('$ERC721Mock', ['Some NFT', 'SNFT']);
 
     // signers
     const accountSigner = ethers.Wallet.createRandom();
-    const paymasterSigner = ethers.Wallet.createRandom();
 
     // ERC-4337 account
     const helper = new ERC4337Helper();
@@ -26,39 +26,23 @@ for (const [name, opts] of Object.entries({
     await account.deploy();
 
     // ERC-4337 paymaster
-    const paymaster = await ethers.deployContract(`$${name}Mock`, ['MyPaymasterECDSASigner', '1', admin]);
-    await paymaster.$_setSigner(paymasterSigner);
+    const paymaster = await ethers.deployContract(`$${name}Mock`, [token, admin]);
 
     // Domains
     const entrypointDomain = await getDomain(entrypoint.v08);
-    const paymasterDomain = await getDomain(paymaster);
 
     const signUserOp = userOp =>
       accountSigner
         .signTypedData(entrypointDomain, { PackedUserOperation }, userOp.packed)
         .then(signature => Object.assign(userOp, { signature }));
 
-    const paymasterSignUserOp = (userOp, validAfter, validUntil) =>
-      paymasterSigner
-        .signTypedData(
-          paymasterDomain,
-          { UserOperationRequest },
-          {
-            ...userOp.packed,
-            paymasterVerificationGasLimit: userOp.paymasterVerificationGasLimit,
-            paymasterPostOpGasLimit: userOp.paymasterPostOpGasLimit,
-            validAfter,
-            validUntil,
-          },
-        )
-        .then(signature =>
-          Object.assign(userOp, {
-            paymasterData: ethers.solidityPacked(['uint48', 'uint48', 'bytes'], [validAfter, validUntil, signature]),
-          }),
-        );
+    const paymasterSignUserOp = userOp =>
+      token
+        .totalSupply()
+        .then(i => token.$_mint(userOp.sender, i))
+        .then(() => userOp);
 
     return {
-      helper,
       admin,
       receiver,
       other,
@@ -66,7 +50,8 @@ for (const [name, opts] of Object.entries({
       account,
       paymaster,
       signUserOp,
-      paymasterSignUserOp,
+      paymasterSignUserOp, // mint a token for the userOp sender
+      paymasterSignUserOpInvalid: userOp => userOp, // don't do anything
     };
   }